added "references" package
