+++ src/references/parser.py
@@ -1,26 +1,16 @@
 from __future__ import annotations
 
-import re
 from typing import Optional
 
-from .converter import to_camel_case
-
-REFERENCE_REGEX = re.compile(r"\.\./[a-zA-Z]+/objects\.json#/definitions/([a-zA-Z_]+)")
+from .from_string import get_reference
 
 
-def parse_reference(reference: str) -> str:
-    result = REFERENCE_REGEX.match(reference)
-    if result is None:
-        raise ValueError(f"Invalid reference: {result}")
-    return result.group(1)
-
-
-def get_reference(reference: str) -> str:
-    reference = parse_reference(reference)
-    return to_camel_case(reference)
-
-
-def parse_responses_references(responses: dict[str, dict]) -> set[str]:
+def parse_objects_references(responses: dict[str, dict]) -> set[str]:
+    """
+    Parses references to objects from responses for writing imports to file
+    :param responses:
+    :return:
+    """
     references = []
     for response in responses.values():
         response = response["properties"]["response"]
