+++ src/strings/__init__.py
@@ -1,13 +1,10 @@
 from .converter import to_camel_case, to_python_type, to_python_types
-from .reference import get_reference, parse_responses_references
 from .validators import is_valid_name, validate_name
 
 __all__ = [
     "to_camel_case",
-    "get_reference",
     "to_python_types",
     "to_python_type",
     "validate_name",
     "is_valid_name",
-    "parse_responses_references",
 ]
