+++ tests/test_strings/test_validators.py
@@ -1,10 +1,6 @@
 import pytest
 
-from src.strings.validators import (
-    is_valid_name,
-    validate_name,
-    validate_order_references,
-)
+from src.strings.validators import is_valid_name, validate_name
 
 
 @pytest.mark.parametrize(
@@ -29,42 +25,3 @@ def test_is_valid_name(string: str, expected: bool):
 )
 def test_validate_name(string: str, expected: str):
     assert validate_name(string) == expected
-
-
-@pytest.mark.parametrize(
-    "objects, expected",
-    [
-        (
-            {
-                "A": {"allOf": [{"$ref": "../dir/objects.json#/definitions/B"}]},
-                "B": {"allOf": [{"$ref": "../dir/objects.json#/definitions/C"}]},
-                "C": {"type": "object"},
-            },
-            {
-                "C": {"type": "object"},
-                "B": {"allOf": [{"$ref": "../dir/objects.json#/definitions/C"}]},
-                "A": {"allOf": [{"$ref": "../dir/objects.json#/definitions/B"}]},
-            },
-        ),
-        (
-            {
-                "A": {"oneOf": [{"$ref": "../dir/objects.json#/definitions/B"}]},
-                "B": {"type": "object"},
-            },
-            {
-                "B": {"type": "object"},
-                "A": {"oneOf": [{"$ref": "../dir/objects.json#/definitions/B"}]},
-            },
-        ),
-        (
-            {"A": {"$ref": "../dir/objects.json#/definitions/B"}, "B": {"type": "object"}},
-            {"B": {"type": "object"}, "A": {"$ref": "../dir/objects.json#/definitions/B"}},
-        ),
-        (
-            {"A": {"type": "object"}, "B": {"type": "object"}},
-            {"A": {"type": "object"}, "B": {"type": "object"}},
-        ),
-    ],
-)
-def test_validate_order_references(objects: dict, expected: dict):
-    assert validate_order_references(objects) == expected
